requirements.txt
ldap3
python-dotenv
streamlit
matplotlib
jira